Created Elevator pitch instructions component
