+++ src/client/components/ElevatorPitchInstructions/ElevatorPitchInstructions.stories.js
@@ -1,0 +1,9 @@
+import React from 'react';
+import ElevatorPitchInstructions from './ElevatorPitchInstructions.component';
+
+export default {
+  title: 'components / Elevator pitch instructions ',
+  component: ElevatorPitchInstructions,
+};
+
+export const Errorpage = () => <ElevatorPitchInstructions />;
