+++ src/client/components/ElevatorPitchInstructions/ElevatorPitchInstructions.styles.css
@@ -1,41 +1,9 @@
-.body-background-image {
-  background-image: url(../../assets/images/Vector.svg);
-  background-size: cover;
-  background-repeat: no-repeat;
-  width: 100%;
-  height: 100vh;
-}
-
-.header h1 {
-  margin-left: 3.4em;
-  letter-spacing: 9px;
-}
-
-.page-content {
-  display: flex;
-}
-
-.sidebar {
-  border: 2px solid yellow;
-  margin: 17rem 0 0 0;
-  min-width: 5%;
-  height: 20%;
-}
-
-.side_menu_div {
-  margin-left: 0;
-  background-color: #5088a8;
-  width: 80%;
-}
-
 .main-content {
-  border: 3px solid green;
   width: 70%;
   margin: 5rem auto;
 }
 
 .top-text-content {
-  border: 2px solid yellow;
   font-size: 1.6rem;
   line-height: normal;
   text-align: center;
@@ -43,19 +11,12 @@
 
 .main-text-content-and-image {
   display: flex;
-  border: 3px solid red;
-  margin-top: 4.5rem;
+  margin-top: 3.5rem;
 }
 
 .left-text-content {
-  width: 61%;
-  border: 2px solid yellow;
-  margin-left: 1.2rem;
-}
-
-.right-image-content {
-  width: 39%;
-  border: 2px solid purple;
+  width: 60%;
+  margin-right: 1rem;
 }
 
 .left-text-content p,
@@ -90,18 +51,61 @@ ul.elevator-pitch-list li:first-of-type {
 .btn-elevator-pitch .buttons {
   font-size: 1.2rem;
   text-align: center;
-  width: 25%;
-  margin: 0.5rem 0;
+  min-width: 25%;
+  margin-top: 3rem;
 }
 
 .top-text-content p:first-of-type {
   margin: 0;
-
-  /* font-weight: 550; */
 }
 
 .top-text-content p:last-of-type {
   margin: 0;
   color: #298f8a;
   font-weight: 600;
+}
+
+.right-image-content {
+  background-color: #d0dddc;
+  width: 40%;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  position: relative;
+  margin-top: 1rem;
+}
+
+.pagination {
+  background-color: #6b748b;
+  text-align: center;
+  color: #f0f0f0;
+  margin: 1rem 0;
+  width: 80%;
+  padding: 0.2rem;
+}
+
+.background-elevator-men {
+  background-color: #f0f0f0;
+  display: flex;
+  justify-content: center;
+  align-items: flex-end;
+  height: 88%;
+  width: 80%;
+}
+
+.background-elevator-men img {
+  width: 65%;
+  height: 85%;
+}
+
+.arrows {
+  display: flex;
+  flex-direction: column;
+  position: absolute;
+  right: 3%;
+}
+
+.arrows img:first-of-type {
+  padding-bottom: 1rem;
 }
